requirements.txt
Jinja2
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
